back_end/requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyserial==3.5
python-barcode==0.16.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
vgamepad==0.1.0
Werkzeug==3.1.3
